requirements.txt
Django==2.1.3
django-oauth-toolkit==1.2.0
djangorestframework==3.9.0
pytz==2018.7